Shop_API/requirement.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
enum34==1.1.10
execnet==2.1.2
flake8==7.3.0
idna==3.11
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mimesis==18.0.0
namedlist==1.8
packaging==25.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
pygments==2.19.2
pytest==8.4.2
pytest-rerunfailures==16.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dotenv==1.2.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
